+++ src/firecrest/filesystem/ops/models.py
@@ -135,7 +135,7 @@ class PostMkdirResponse(CamelModel):
 
 class PostCompressRequest(FilesystemRequestBase):
     target_path: str
-    pattern: Optional[str] = None
+    match_pattern: Optional[str] = None
     dereference: Optional[bool] = False