+++ src/firecrest/filesystem/transfer/models.py
@@ -73,7 +73,7 @@ class MoveResponse(CamelModel):
 class CompressRequest(FilesystemRequestBase):
     target_path: str
     account: Optional[str] = None
-    pattern: Optional[str] = None
+    match_pattern: Optional[str] = None
     dereference: Optional[bool] = False