+++ tests/filesystem_transfer_test.py
@@ -358,6 +358,40 @@ async def test_compress(
         assert compress.transfer_job.job_id == mocked_job_submit_response["job_id"]
         assert compress.transfer_job.system == slurm_cluster_with_api_config.name
 
+@pytest.mark.asyncio
+async def test_compress_with_pattern(
+    client,
+    slurm_cluster_with_api_config,
+    mocked_job_submit_response,
+):
+
+    request_body = {
+        "sourcePath": "/home/test1/file1.txt",
+        "targetPath": "/home/test1/file1.tar",
+        "account": "fireuser",
+        "dereference": False,
+        "match_pattern": "./[ab].*\\.txt"
+    }
+
+    with aioresponses() as mocked:
+        mocked.post(
+            f"{slurm_cluster_with_api_config.scheduler.api_url}/slurm/v{slurm_cluster_with_api_config.scheduler.api_version}/job/submit",
+            status=200,
+            body=json.dumps(mocked_job_submit_response),
+        )
+
+        # Testing storage end-point
+        response = client.post(
+            f"/filesystem/{slurm_cluster_with_api_config.name}/transfer/compress",
+            json=request_body,
+        )
+        assert response.status_code == 201
+        assert response.json() is not None
+        compress = CompressResponse(**response.json())
+
+        assert compress.transfer_job.job_id == mocked_job_submit_response["job_id"]
+        assert compress.transfer_job.system == slurm_cluster_with_api_config.name
+
 
 @pytest.mark.asyncio
 async def test_extract(